requirements.txt
# Core ML
tensorflow==2.13.0
numpy==1.24.3
h5py==3.8.0

# Data processing
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1

# Image handling
Pillow==9.5.0

# App
streamlit==1.32.0